+++ src/mpservice/_http.py
@@ -17,9 +17,12 @@ from mpservice.multiprocessing import MP_SPAWN_CTX, SpawnProcess
 logger = logging.getLogger(__name__)
 
 
+# See `uvicorn`.
 class Server(uvicorn.Server):
     def run(
-        self, sockets: list[socket.socket] | None = None, worker_idx: int = 0
+        self, stop_requested: multiprocessing.synchronize.Event,
+        sockets: list[socket.socket] | None = None, 
+        worker_idx: int = 0
     ) -> None:
         # When there are multiple worker processes, this method is the
         # `target` function that runs in a new process.
@@ -29,7 +32,8 @@ class Server(uvicorn.Server):
         os.environ['UVICORN_WORKER_IDX'] = str(worker_idx)
         global _stop_requested
         assert _stop_requested is None, f"{_stop_requested} is None"
-        _stop_requested = self._stop_requested
+        _stop_requested = stop_requested
+        self._stop_requested = stop_requested  # to be used in `on_tick`.
 
         return super().run(sockets)
 
@@ -45,14 +49,14 @@ class Server(uvicorn.Server):
 # See `uvicorn`.
 class Multiprocess(uvicorn.supervisors.Multiprocess):
     def run(self, stop_requested: multiprocessing.synchronize.Event):
-        self.startup()
+        self.startup(stop_requested)
         while True:
             if self.should_exit.is_set() or stop_requested.is_set():
                 break
             time.sleep(0.15)
         self.shutdown()
 
-    def startup(self) -> None:
+    def startup(self, stop_requested: multiprocessing.synchronize.Event) -> None:
         message = "Started parent process [{}]".format(str(self.pid))
         color_message = "Started parent process [{}]".format(
             click.style(str(self.pid), fg="cyan", bold=True)
@@ -68,7 +72,8 @@ class Multiprocess(uvicorn.supervisors.Multiprocess):
                 target=self.target,
                 sockets=self.sockets,
                 worker_idx=_idx,
-                # Customization: pass in `worker_idx`
+                stop_requested=stop_requested,
+                # Customization: pass in `worker_idx` and `stop_requested`.
             )
             process.start()
             self.processes.append(process)
@@ -80,6 +85,7 @@ def get_subprocess(
     target: Callable[..., None],
     sockets: List[socket.socket],
     worker_idx: int,
+    stop_requested: multiprocessing.synchronize.Event,
 ) -> SpawnProcess:
     """
     Called in the parent process, to instantiate a new child process instance.
@@ -94,9 +100,10 @@ def get_subprocess(
     # We pass across the stdin fileno, and reopen it in the child process.
     # This is required for some debugging environments.
 
-    # There are two customizations to this function:
+    # There are three customizations to this function:
     #   - take ``worker_idx``
     #   - use ``MP_SPAWN_CTX``
+    #   - use ``stop_requested``
 
     stdin_fileno: Optional[int]
     try:
@@ -110,6 +117,7 @@ def get_subprocess(
         "sockets": sockets,
         "stdin_fileno": stdin_fileno,
         "worker_idx": worker_idx,
+        "stop_requested": stop_requested,
     }
 
     return MP_SPAWN_CTX.Process(target=subprocess_started, kwargs=kwargs)
@@ -122,6 +130,7 @@ def subprocess_started(
     sockets: List[socket.socket],
     stdin_fileno: Optional[int],
     worker_idx: int,
+    stop_requested: multiprocessing.synchronize.Event,
 ) -> None:
     """
     Called when the child process starts.
@@ -145,10 +154,14 @@ def subprocess_started(
     config.configure_logging()
 
     # Now we can call into `Server.run(sockets=sockets)`
-    target(sockets=sockets, worker_idx=worker_idx)
+    target(stop_requested=stop_requested, sockets=sockets, worker_idx=worker_idx)
 
 
 _stop_requested: MP_SPAWN_CTX.Event() = None
+# When a new process is spawned, this variable will be `None`
+# in that process upon importing this module.
+# This variable is initialized in `start_server` in the "main" process
+# and re-assigned by `Server.run` in the main and child processes.
 
 
 async def stop_server():
@@ -171,7 +184,7 @@ def start_server(
     **kwargs,
 ) -> uvicorn.Server:
     """
-    This function is specifically for use with :mod:`mpservice.mpserver`.
+    This function is intended for lauching a service that uses :mod:`mpservice.mpserver`.
 
     This function is adapted from ``uvicorn.main.run``.
 
@@ -186,9 +199,10 @@ def start_server(
             python example.py
 
         in the directory that contains `example.py`, then Python will be able to import that script.
-        In this case, `app` can be written as `'example:app'`.
+        In that case, `app` can be written as `'example:app'`.
 
-        If `workers > 1`, this must be a str.
+        If `workers > 1`, this must be a str, because a `Starlette` object cannot be pickled and sent
+        to other processes.
     host
         The default ``'0.0.0.0'`` is suitable if the code runs within a Docker container.
     port
@@ -207,8 +221,16 @@ def start_server(
         ``uvicorn`` uses a large default value (2048). That might be misguided.
 
         See: https://bugs.python.org/issue38699, https://stackoverflow.com/a/2444491, https://stackoverflow.com/a/2444483,
+    workers
+        If ``1``, the server is launched in the current process/thread. If ``> 1``, this many processes
+        are spawned, each running one copy of ``app`` independently.
+
+        Note that worker processes are not dynamically created and killed according to need---this fixed
+        number of processes are started and they remain active---this is in contrast to Gunicorn.
+        This also makes our customization---placing ``UVICORN_WORKER_IDX`` in the environ---meaningful.
     **kwargs
         Passed to ``uvicorn.Config``.
+
         If user has their own ways to config logging, then pass in
         ``log_config=None`` through ``**kwargs``.
     """
@@ -223,13 +245,13 @@ def start_server(
     )
 
     server = Server(config=config)
-    server._stop_requested = MP_SPAWN_CTX.Event()
+    stop_requested = MP_SPAWN_CTX.Event()
     if workers == 1:
-        server.run()
+        server.run(stop_requested)
     else:
         sock = config.bind_socket()
         Multiprocess(config, target=server.run, sockets=[sock]).run(
-            server._stop_requested
+            stop_requested
         )
 
     if config.uds and os.path.exists(config.uds):