+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.14.1] - in progress
 
-- New http serving code in ``mpservice.experimental.http`` designed for multiple worker processes
+- New http serving code in ``mpservice.http`` designed for multiple worker processes
   managed by ``uvicorn`` (as opposed to multiple worker processes managed by ``mpservice.mpserver``).